Support Latin-1 encoded data
Gopher servers are expected to send their data encoded as ASCII. But nowadays
UTF-8 is basically everywhere. And as UTF-8 is a superset of ASCII, gopish uses
UTF-8 for all its encoding needs.

That works fine, as long as the server does not use another, incompatible ASCII
superset as encoding. Probably the most widespread incompatible ASCII superset
is Latin-1.

gophish now tries to decode received data as UTF-8 and if that operation fails,
Latin-1 is used.

For encoding purposes (especially selectors sent to servers) UTF-8 is used.