Replace BusyIndicator with ProgressBar
Transferring data using the gopher protocol is fast, but rendering large gopher
menus is a bottleneck and can take a significant amount of time. To give the
user a better feeling for the progress, replace the current BusyIndicators with
ProgressBars and, in the case of gopher menus, display the real progress values.

In the case of gopher menus, the ProgressBar is indeterminate until the data has
been received and parsed. Once the rendering starts, the ProgressBar represents
the real progress.